chore: Rename archive package to archivetest
